added checks
