--- cypress/e2e/data-import/settings/create-match-profile-for-marc-bib-matching-to-folio-record-type.cy.js
@@ -1,44 +1,0 @@
-import getRandomPostfix from '../../../support/utils/stringTools';
-import { DevTeams, Permissions, TestTypes } from '../../../support/dictionary';
-import { EXISTING_RECORDS_NAMES } from '../../../support/constants';
-import SettingsMenu from '../../../support/fragments/settingsMenu';
-import Users from '../../../support/fragments/users/users';
-import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
-import MatchProfileView from '../../../support/fragments/data_import/match_profiles/matchProfileView';
-import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
-
-describe('data-import', () => {
-  describe('Settings', () => {
-    let user;
-    const matchProfile = {
-      profileName: `C368009 001 to Instance HRID ${getRandomPostfix()}`,
-      incomingRecordFields: {
-        field: '001',
-      },
-      matchCriterion: 'Exactly matches',
-      existingRecordType: EXISTING_RECORDS_NAMES.INSTANCE,
-      instanceOption: NewMatchProfile.optionsList.instanceHrid,
-    };
-
-    before('create user', () => {
-      cy.createTempUser([Permissions.settingsDataImportEnabled.gui]).then((userProperties) => {
-        user = userProperties;
-        cy.login(user.username, user.password);
-      });
-    });
-
-    it(
-      'C9321 Create match profile for MARC Bib matching to a FOLIO record type (folijet) (TaaS)',
-      { tags: [TestTypes.extendedPath, DevTeams.folijet] },
-      () => {
-        cy.visit(SettingsMenu.matchProfilePath);
-        MatchProfiles.verifyListOfExistingProfilesIsDisplayed();
-        MatchProfiles.openNewMatchProfileForm();
-        NewMatchProfile.fillName(matchProfile.profileName);
-        NewMatchProfile.verifyExistingRecordSection();
-        NewMatchProfile.selectExistingRecordType(matchProfile.existingRecordType);
-        NewMatchProfile.verifyExistingRecordTypeIsSelected(matchProfile.existingRecordType);
-      },
-    );
-  });
-});
